refactor azdo code to multiple files in azdo package
